+++ app.py
@@ -12,6 +12,7 @@ from flask import Flask, request, Response
 import jsonpickle
 import cv2
 import numpy as np
+import uuid
 import base64
 import os
 # Initialize the Flask application
@@ -23,10 +24,10 @@ config = {'orient' : True,    #corrects orientation of image default -> True
           'contrast' : True,  #Bnw for Better OCR default -> True
           'psm': [3,4,6],     #Google Tesseract  psm modes by default 3,4,6. 
           'mask_color': (0, 165, 255),  #Masking color BGR Format
-          'brut_psm': [6]     #Note : Keep only one psm for brut mask (6) is good to start
+          'brut_psm': [3,4,6]     #Note : Keep only one psm for brut mask (6) is good to start
           }
 
-ac = Aadhaar_Card(config)
+
 
 def to_image_string(image_filepath):
     return base64.b64encode(open(image_filepath, 'rb').read())#.encode('base64')
@@ -38,60 +39,77 @@ def delete_file(path):
 #  else:
 #    print("Sorry, file %s does not exist." % path)
 
-# validate method to check sequence
-@app.route('/api/validate', methods=['GET','POST'])
-def validate():
-    r = request.get_json(force=True)   #content type application/json
-    #print("this s",type(r))
-    #print(r)
-    content_type = 'application/json'
-    headers = {'content-type': content_type}
-    
-    return Response(response=jsonpickle.encode({'validity':ac.validate(str(r['test_number']))}), status=200, mimetype="application/json", headers=headers)
 
+
+#-----------------------------------[1]  Index.html for GUI-----------------------------------------------------
 def get_file(filename):  # pragma: no cover
     try:
         src = os.path.join(filename)
-        # Figure out how flask returns static files
-        # Tried:
-        # - render_template
-        # - send_file
-        # This should not be so non-obvious
         return open(src).read()
     except IOError as exc:
         return str(exc)
 
 @app.route('/', methods=['GET'])
-def metrics():  # pragma: no cover
+def indexfile():  # pragma: no cover
     content = get_file('public/imagebase64.html')
     return Response(content, mimetype="text/html")
+#---------------------------------------------------------------------------------------
 
-#Extract Aadhaar from a base64 image
+
+
+#------------------------------------[2] Validate method to check sequence in aadhar Number-----------------------------------------------
+@app.route('/api/validate', methods=['GET','POST'])
+def validate():
+    ac = Aadhaar_Card(config);
+    r = request.get_json(force=True)   #content type application/json
+    content_type = 'application/json'
+    headers = {'content-type': content_type}
+    ac = Aadhaar_Card(config);
+    validity=ac.validate(str(r['test_number']))
+    
+    return Response(response=jsonpickle.encode({'validity':validity}), status=200, mimetype="application/json", headers=headers)
+#-----------------------------------------------------------------------------------------------------------------------------------------
+
+
+#------------------------------------[3] Extract Aadhaar from a base64 image--------------------------------------------------------------
 @app.route('/api/ocr', methods=['GET','POST'])
 def ocr():
-#    print("request",request)
+    ac = Aadhaar_Card(config);
+    requestid=uuid.uuid4();
+    temp_name = "input/"+str(requestid)+".png" 
     r = request.get_json(force=True)  #force=True #content type application/json
-    temp_name = "123.png" 
+    
     image = r['doc_b64'] # raw data with base64 encoding
+    
     decoded_data = base64.b64decode(image)
     np_data = np.fromstring(decoded_data,np.uint8)
+    
     img = cv2.imdecode(np_data,cv2.IMREAD_UNCHANGED)
+    cv2.imwrite(temp_name,img)
     content_type = 'application/json'
     headers = {'content-type': content_type}
     try:
-        print("Extracting Addhar")
-        print(ac.extract(temp_name))
-    except:
-        return Response(response=jsonpickle.encode({'aadhaar_list':'ac.extract(temp_name)'}), status=200, mimetype="application/json", headers=headers)
+        #print("Extracting Addhar")
+        print(ac.uuid);
+        aadhar=ac.extract(temp_name)
+        delete_file(temp_name)
+        return Response(response=jsonpickle.encode({'aadhaar_list':aadhar}), status=200, mimetype="application/json", headers=headers)
+    except Exception as e:
+        print(e)
+        return Response(response=jsonpickle.encode({'aadhaar_list':''}), status=200, mimetype="application/json", headers=headers)
+#-----------------------------------------------------------------------------------------------------------------------------------------
 
 
-#Mask aadhaar number card for given aadhaar card number
 
+#------------------------------------[4] Mask aadhaar number card for given aadhaar card number-------------------------------------------
 @app.route('/api/mask', methods=['GET','POST'])
 def mask():
+    ac = Aadhaar_Card(config);
     flag_mask = 0
+    requestid=uuid.uuid4();
+    temp_name = "input/"+str(requestid)+".png" 
     r = request.get_json(force=True)  #force=True #content type application/json
-    temp_name = "temp_unmasked.png" 
+    #temp_name = "temp_unmasked.png" 
     image = r['doc_b64'] # raw data with base64 encoding
     decoded_data = base64.b64decode(image)
     np_data = np.fromstring(decoded_data,np.uint8)
@@ -103,7 +121,7 @@ def mask():
     delete_file(temp_name)
     content_type = 'application/json'
     headers = {'content-type': content_type}
-#    print(flag_mask)
+
     if flag_mask == 0:
         return Response(response=jsonpickle.encode({'doc_b64_masked':'None', 'is_masked': False}), status=200, mimetype="application/json", headers=headers)
     else:
@@ -112,11 +130,13 @@ def mask():
         #convert byte to string
         encoded_string = img_bytes.decode("utf-8")
         return Response(response=jsonpickle.encode({'doc_b64_masked':encoded_string, 'is_masked': True}), status=200, mimetype="application/json", headers=headers)
+#-----------------------------------------------------------------------------------------------------------------------------------------
+
 
 #Brut Mask any Readable Number from Aadhaar (works good for low res and bad quality images)
-
 @app.route('/api/brut_mask', methods=['GET','POST'])
 def brut_mask():
+    ac = Aadhaar_Card(config);
     r = request.get_json(force=True)  #force=True #content type application/json
     temp_name = "temp_unmasked.png" 
     image = r['doc_b64'] # raw data with base64 encoding
@@ -142,8 +162,9 @@ def brut_mask():
 #Process : Image -> Extract Text -> Check for aadhaar number -> Mask first 8 digits // check validity of aadhaar number optional
 #          If aadhaar card number is not found using OCR, try brut mode and mask possible numbers.
 
-@app.route('/api/sample_pipe', methods=['GET','POST'])
+@app.route('/api/Mask8Digits', methods=['GET','POST'])
 def sample_pipe():
+    ac = Aadhaar_Card(config);
     flag_mask = 0
     r = request.get_json(force=True)  #force=True #content type application/json
     temp_name = "temp_unmasked.png"
